+++ src/main/java/items/mythicalcreatures/Dragon.java
@@ -2,7 +2,7 @@ package items.mythicalcreatures;
 
 public class Dragon extends MythicalCreature{
 
-    public Dragon(String name, int defenceMultiplier) {
+    public Dragon(String name, double defenceMultiplier) {
         super(name, defenceMultiplier);
     }
 }
