+++ src/main/java/items/mythicalcreatures/MythicalCreature.java
@@ -3,14 +3,16 @@ package items.mythicalcreatures;
 public abstract class MythicalCreature {
 
     protected String name;
-    protected int defenceMultiplier;
+    protected double defenceMultiplier;
 
-    public MythicalCreature(String name, int defenceMultiplier) {
+    public MythicalCreature(String name, double defenceMultiplier) {
         this.name = name;
         this.defenceMultiplier = defenceMultiplier;
     }
 
-    public int getDefenceMultiplier() {
+    public double getDefenceMultiplier() {
         return defenceMultiplier;
     }
+
+
 }
