+++ src/main/java/items/weapons/Sword.java
@@ -4,7 +4,7 @@ import items.AttackingItem;
 
 public class Sword extends AttackingItem {
 
-    public Sword(String name, int attackMultiplier) {
+    public Sword(String name, double attackMultiplier) {
         super(name, attackMultiplier);
     }
 }
